Move packages to its own folder
