+++ com.unity.mobile.android-logcat/Editor/UI/Packages.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 598088db5abdbee4ab7e54a093b55c5c
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
