+++ com.unity.mobile.android-logcat/Editor/UI/Packages/AndroidLogcatPackageProperties.cs.meta
